requirements.txt
numpy~=1.24.2
opencv-python~=4.7.0.72
opencv-contrib-python~=4.7.0.72
Image~=1.5.33
Pillow~=9.5.0